Change ROC param to uint32

+++ rtp-session.go
@@ -65,9 +65,9 @@ func (s *RTPSession) Decode(packetData []byte) (*RTPPacket, error) {
 
 	if s.cipherID == HALF_AEAD_AES_128_GCM_AEAD_AES_128_GCM {
 		cipherKeySize := 128 / 8
-		cipherKeyEnc := s.kdf.Derive(Ke, uint64(s.roc), seq, cipherKeySize)
+		cipherKeyEnc := s.kdf.Derive(Ke, s.roc, seq, cipherKeySize)
 		//cipherKeyAuth := s.kdf.Derive(Ka, s.roc, s.seq, cipherKeySize) // not used for GCM
-		cipherSalt := s.kdf.Derive(Ks, uint64(s.roc), seq, cipherKeySize)
+		cipherSalt := s.kdf.Derive(Ks, s.roc, seq, cipherKeySize)
 
 		err := p.DecryptGCM(s.roc, cipherKeyEnc, cipherSalt)
 		if err != nil {
@@ -102,9 +102,9 @@ func (s *RTPSession) Encode(p *RTPPacket) ([]byte, error) {
 
 		// encrypt
 		cipherKeySize := 128 / 8
-		cipherKeyEnc := s.kdf.Derive(Ke, uint64(s.roc), s.seq, cipherKeySize)
+		cipherKeyEnc := s.kdf.Derive(Ke, s.roc, s.seq, cipherKeySize)
 		//cipherKeyAuth := s.kdf.Derive(Ka, s.roc, s.seq, cipherKeySize) // not used for GCM
-		cipherSalt := s.kdf.Derive(Ks, uint64(s.roc), s.seq, cipherKeySize)
+		cipherSalt := s.kdf.Derive(Ks, s.roc, s.seq, cipherKeySize)
 
 		err = p.EncryptGCM(s.roc, cipherKeyEnc, cipherSalt)
 		if err != nil {
